Create DepartmentService
Implemented DepartmentServiceImpl
Refactored EmployeeService with replacing int parameter to Long

Closes #4

+++ src/main/java/com/arsenbaktiyarov/spring/springboot/springjpa/repository/EmployeeRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EmployeeRepository extends CrudRepository<Employee,Integer> {
+public interface EmployeeRepository extends CrudRepository<Employee,Long> {
     public List<Employee> findAllByName(String name);
 
 }
